Removed need for forward declaration, hopefully this will make it compile on gcc
